Split `to_xarray` off into its own file
